requirements.txt
mujoco
mink
hydra-core